+++ vice/src/video/render1x1pal.c
@@ -67,7 +67,8 @@ void render_16_1x1_pal(const DWORD *colortab, const BYTE *src, BYTE *trg,
         wfast = 0;
         wend = 0;
     } else {
-        wstart = 8 - ((unsigned int)trg & 7); /* alignment: 8 pixels*/
+        /* alignment: 8 pixels*/
+        wstart = (unsigned int)(8 - ((unsigned long)trg & 7));
         wfast = (width - wstart) >> 3; /* fast loop for 8 pixel segments*/
         wend = (width - wstart) & 0x07; /* do not forget the rest*/
     }
@@ -161,7 +162,8 @@ void render_32_1x1_pal(const DWORD *colortab, const BYTE *src, BYTE *trg,
         wfast = 0;
         wend = 0;
     } else {
-        wstart = 8 - ((unsigned int)trg & 7); /* alignment: 8 pixels*/
+        /* alignment: 8 pixels*/
+        wstart = (unsigned int)(8 - ((unsigned long)trg & 7));
         wfast = (width - wstart) >> 3; /* fast loop for 8 pixel segments*/
         wend = (width - wstart) & 0x07; /* do not forget the rest*/
     }
