+++ vice/src/video/render1x2.c
@@ -52,7 +52,8 @@ void render_08_1x2_04(const DWORD *colortab, const BYTE *src, BYTE *trg,
         wfast = 0;
         wend = 0;
     } else {
-        wstart = 8 - ((unsigned int)trg & 7); /* alignment: 8 pixels*/
+        /* alignment: 8 pixels*/
+        wstart = (unsigned int)(8 - ((unsigned long)trg & 7));
         wfast = (width - wstart) >> 3; /* fast loop for 8 pixel segments*/
         wend = (width - wstart) & 0x07; /* do not forget the rest*/
     }
@@ -125,7 +126,8 @@ void render_16_1x2_04(const DWORD *colortab, const BYTE *src, BYTE *trg,
         wfast = 0;
         wend = 0;
     } else {
-        wstart = 8 - ((unsigned int)trg & 7); /* alignment: 8 pixels*/
+        /* alignment: 8 pixels*/
+        wstart = (unsigned int)(8 - ((unsigned long)trg & 7));
         wfast = (width - wstart) >> 3; /* fast loop for 8 pixel segments*/
         wend  = (width - wstart) & 0x07; /* do not forget the rest*/
     }
@@ -199,7 +201,8 @@ void render_24_1x2_04(const DWORD *colortab, const BYTE *src, BYTE *trg,
         wfast = 0;
         wend = 0;
     } else {
-        wstart = 4 - ((unsigned int)trg & 3); /* alignment: 4 pixels*/
+        /* alignment: 4 pixels*/
+        wstart = (unsigned int)(4 - ((unsigned long)trg & 3));
         wfast = (width - wstart) >> 2; /* fast loop for 4 pixel segments*/
         wend = (width - wstart) & 0x03; /* do not forget the rest*/
     }
@@ -328,7 +331,8 @@ void render_32_1x2_04(const DWORD *colortab, const BYTE *src, BYTE *trg,
         wfast = 0;
         wend = 0;
     } else {
-        wstart = 8 - ((unsigned int)trg & 7); /* alignment: 8 pixels*/
+        /* alignment: 8 pixels*/
+        wstart = (unsigned int)8 - ((unsigned long)trg & 7);
         wfast = (width - wstart) >> 3; /* fast loop for 8 pixel segments*/
         wend  = (width - wstart) & 0x07; /* do not forget the rest*/
     }
